ArticleSearchParameter renamed to ArticleLoadParameter
