+++ component/src/main/java/com/celements/blog/article/IArticleEngineRole.java
@@ -7,7 +7,7 @@ import org.xwiki.component.annotation.ComponentRole;
 @ComponentRole
 public interface IArticleEngineRole {
   
-  public List<Article> getArticles(ArticleSearchParameter param
+  public List<Article> getArticles(ArticleLoadParameter param
       ) throws ArticleLoadException;
 
 }
